requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
email-validator>=2.2.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pypdf>=5.4.0
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
striprtf>=0.0.26
pytesseract>=0.3.10
PyMuPDF>=1.24.0
extract-msg>=0.52.0
EbookLib>=0.18
beautifulsoup4>=4.12.0
mobi>=0.4.1
fpdf2>=2.8.0
httpx>=0.28.0
cryptography>=43.0.0
pgvector>=0.3.6
redis>=5.0.0
slowapi>=0.1.9
python-json-logger>=2.0.7
stripe>=11.0.0,<16.0.0
